requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
black==25.1.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.0
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.58.0
frozenlist==1.6.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
grpclib==0.4.7
h2==4.2.0
h5py==3.13.0
hpack==4.1.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.0
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.8.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
modal==1.0.1
multidict==6.4.4
mypy_extensions==1.1.0
namex==0.0.9
nest-asyncio==1.6.0
numpy==2.0.2
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.2.1
pillow_heif==0.22.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==26.4.0
requests==2.32.3
rich==14.0.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
setuptools==80.7.1
shellingham==1.5.4
sigtools==4.0.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
synchronicity==0.9.12
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
threadpoolctl==3.6.0
toml==0.10.2
tomli==2.2.1
tornado==6.4.2
traitlets==5.14.3
typer==0.15.4
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchfiles==1.0.5
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
yarl==1.20.0
zipp==3.21.0
